Export ConfigUpdateWatcher struct + update doc comments
Summary:
The `ConfigUpdateWatcher` struct in the `cached_config` crate is meant to be directly usable by callers. This diff exports that struct.

The usefulness of `ConfigUpdateWatcher` is tied to the parent `ConfigHandle`, i.e. as long as the corresponding `ConfigHandle` exists, the associated `ConfigUpdateWatchers` will continue to get updated configurations. Once the `ConfigHandle` is dropped, the watchers will no longer be notified of changes. This information was not reflected in the doc comments and has now been added.

Differential Revision: D40297199

fbshipit-source-id: 11e844c5f3025b161b67504fe0087d3344205fd3