Fix queries! macro on pub write queries
Summary:
This macro doesn't work on write queries that don't have `values: ` but do have visibility modifiers. It is missing some of the fields which causes errors like `missing ","`.

So, currently
```
write TestQuery7(x: i64) {
    none,
    "REPLACE INTO foo (ID, x) VALUES (1, {x})"
}
```
works perfectly, but
```
pub(crate) write TestQuery7(x: i64) {
    none,
    "REPLACE INTO foo (ID, x) VALUES (1, {x})"
}
```
crashes.

This diff fixes that issue by fixing the macro rules case.

 ---
Sidenote: I was reading the macro, and wanted to improve it by using `:vis` to account for visibility stuff and simplify the code. However, I can't really do that without changing the behaviour slightly.

Because currently if no visibility is specified, it makes the mod private but the functions `pub(super)`, and if `vis` is specified, it makes both the mod and the functions have `vis` visibility.

This is kind of weird, and not possible to simulate using `:vis` type. I could only use `vis` for the `mod` and `pub` for the functions inside, which for me makes sense, and works kind of the same, but works a little differently for re-exports (which seems fine... but does change some behaviour).

What do people think?

Reviewed By: jsgf

Differential Revision: D40181862

fbshipit-source-id: bd882dad14c1e76d17c3328bd9cca61480fe2fbe