Simplify macro visibility rules
Summary:
Currently if you create a query without any visibilty modifier, it creates a **private** module with **pub(super)** functions inside, and if you specify a visibility, it uses that visibilty on both module and functions. The reason it's like this is to avoid creating private functions that even the declarer can't use :).

But in order to do this, the macro needs duplicated rules for "with visibility specified" vs "without visibility specified". By simplifying this, I was able to remove about half of the code for the macro.

The difference is: Now you specify the visibility of the module, but the visibility of the inner functions is always `pub`. This works *slightly* different, but shouldn't break anything.

Reviewed By: jsgf

Differential Revision:
D40509025

LaMa Project: L1122763

fbshipit-source-id: 514aec0faf8d70efae39f4a6c39831d4a9f872ef