Add type parameter (generics) support to `trait_alias` crate
Summary:
As part of my attempt to use this in D40870195 (abandoned for reasons mentioned in that diff), I extended the `trait_alias` macro to support type parameters.

**I no longer have use for this, but I'm publishing the diff just in case anyone else might want it.**

There is also some prior art for this in the OSS community, and at some point, it could make sense to just switch to that: https://github.com/popzxc/trait-set (but there is some value in reducing third-party dependencies).

Reviewed By: markbt

Differential Revision: D40825641

fbshipit-source-id: 3299a3b74f419d4d20c76a6ec599c7474c7bef51