Remove eventually
Summary:
`eventually` is a crate that depends on tokio 0.2. While the author has been aware of tokio 1.0, they have opted into releasing a 0.5 that will update the dependency instead... last year.

The only user of `eventually` right now is `aida`, which has been disabled and is unlikely to be restarted in the future, and whose codebase is for reference only (verified with mehr). Instead of relying on a git revision, it makes more sense to comment out the dependency to remove `eventually` entirely, ensuring we don't run into a runtime hazard.

Reviewed By: chriskonstad

Differential Revision: D41590502

fbshipit-source-id: fd75b9a9710dd41c2b964be7d918d6bb5c295341