change internal impl module name to fix a proc-macro expanding issue
Summary:
Raw litterals are weirdly not supported in the std proc-macro bridge because it nevers honors the is_raw attribute.

A fix in stdlib would take much longer than fixing a name thing in fbinit, so while we make this happen, let's make fbinit not break proc-macro support for rust-analyzer

The issue *might* (totally not secure of this) also spill to other crates since fbinit uses parse_quote which panics, and I suspect this kills all the proc-macro evaluations, since most of them are broken if you depend on fbinit (e.g clap).

Reviewed By: davidbarsky

Differential Revision: D42171445

fbshipit-source-id: af22e3094119522b6b75272d6c25612c18383c55