wal: Use READ COMMITTED isolation level
Summary:
This should prevent us from having a lot of "deadlock" issues on our queries.

This improves our lib to be able to use this isolation level, but that's only used on WAL XDB. I'm not sure what's the implication of using it elsewhere. Probably not bad, but we're not having issues on those I believe.

See https://fb.workplace.com/groups/mysql.users/posts/24130604319894856

Reviewed By: mitrandir77

Differential Revision: D42499929

fbshipit-source-id: c5e784ce4c5d82e1a85c3587b927371ff59d5a12