Fix suspicious_double_ref_op in impl slog::KV for SharedError
Summary:
```
error: using `.deref()` on a double reference, which returns `&dyn std::error::Error` instead of dereferencing the inner type
   --> shed/shared_error/src/anyhow.rs:108:23
    |
108 |         let err = self.deref() as &dyn StdError;
    |                       ^^^^^^^^
    |
note: the lint level is defined here
   --> shed/shared_error/src/lib.rs:11:9
    |
11  | #![deny(warnings, missing_docs, clippy::all, rustdoc::broken_intra_doc_links)]
    |         ^^^^^^^^
    = note: `#[deny(suspicious_double_ref_op)]` implied by `#[deny(warnings)]`

error: could not compile `shared_error` (lib) due to previous error
```

There is no Deref impl for SharedError, so `self.deref()` resolves to the Deref impl on &SharedError, i.e. `Deref::deref(&self)` which is equivalent to `self`. So this is displaying `self as &dyn StdError`, which is exactly the same Display impl as displaying `self`, 2 lines above.

This code is nearly 2 years old ({D31171128}). I can't tell for sure what it was trying to accomplish. But the thing being logged as "root_cause" is identical to "error". Causes are already included in the log for "error_debug".

Reviewed By: diliop

Differential Revision: D47123172

fbshipit-source-id: 4f632f9523c7afd1f908ddd3474733e3b5c1c27a