rust/scuba_sample: replace bool with ShouldLog enum
Summary:
# This Diff
Improves the expressiveness with which we handle `scuba_sample::sampling::Sampling` results, per Pedro's suggestion [here](https://www.internalfb.com/diff/D48807058?transaction_fbid=989674378948222). Previously, it was rather hard to tell whether the `bool` meant:

*"yes, this sample has just been evaluated by a sampling policy and was selected"*
or
*"yes, this sample is eligible to be sent to scuba"*

These may look alike, but the `Sampling` API decouples when a sample is evaluated by a sampling policy versus when it is handled by logging.

As a result, it was quite easy to write (incorrect) code like this:
```
if Sampling::SampledIn(NonZeroU64::new(100).unwrap()).apply(my_sample) {
    // we assume the sample got a lucky 1:100 selection
    // ...
    // but in fact this branch will always be exercised
} else {
    // and this branch is never, ever reached!
}
```

While this change doesn't completely stop us from authoring the above code, it should hopefully make clearer what the results of `apply()` actually mean: that the (already evaluated) sampling policy chose, at some point in the past, that this sample should be / should not be sent to Scuba.

Reviewed By: Imxset21

Differential Revision: D49021403

fbshipit-source-id: afd0656f95b368535c9be87a88764a54f5c74a36