Add ConvertToTuple trait and impl for HList and tuple
Summary:
This diff adds a ConvertToTuple trait and implementations for tuples of N length and Hlist of N length. This way `.convert()` can be called for either type, with both returning a tuple of N length.

This will be useful when processing `mysql_common` query results, as they are either tuples for results with up to 12 coluns but as an HList beyond that.

Differential Revision: D50737241

fbshipit-source-id: 0e93b8da9218fd70440b986d4f6a1b9262ada9be