hlist_to_tuple: Simplify to tuple conversion macro
Summary:
We don't really need the `ToTuple` `struct` that implements `FnOnce`.

That was a first attempt at generating the code, but since we switched to implementing a `ConvertToTuple` trait for both tuples and hlists up-to a certain cardinality, we can inline the implementation from the `FnOnce` and delete it.

We also get to delete a test that only tested that intermediate stage. `test_convert` covers all the functionality we need.

Reviewed By: mitrandir77

Differential Revision: D50738994

fbshipit-source-id: 71d171f321b4ae5d42c8c92d195ef572c87ba2b2