hlist_to_tuple: Intentionally format the macro call
Summary:
With the auto-formatted list, types don't line up, making it harder to verify correctness and to go from 40 to a larger number if we need to.
Line things up in base 10 so it's easier to read and edit.

Skip rustfmt for this line so we can keep our little snowflake shape :)

Reviewed By: clara-9

Differential Revision: D50740330

fbshipit-source-id: 8120780de12b38b4749adb9645395f50f147e729