Move WriteResult to mod
Summary:
Up until this diff the implementation of WriteResult was in the `fbmysql_wrapper`. However the implementation can be used for any mysql implemntation as it is very simple.
Here I extract it to the `mod.rs` file so I can reuse it for the external mysql version later in the stack

Reviewed By: gustavoavena

Differential Revision: D50838010

fbshipit-source-id: 8e9246b116634fbae42c65737482f0ab864650b4