External MySQL abstraction
Summary:
The rust/shed/sql crate provides a wrapper for C++ bindings of the Meta internal MySQL client. This client will not be coompatible with AWS (at least this compatibility will not be supported by MySQL team) and we won't be able to control network dependencies in it.

For the AWS Mononoke instance I'm creating a new abstraction that will be able to speak MySQL with any host and port. Since this new implementation will becloser to the OSS version, I am renaming the internal Meta MySQL client bining to `FBMySQL/FBConnection...`.

Reviewed By: mitrandir77

Differential Revision: D49962569

fbshipit-source-id: 096009b3f0427fe5c382f60341460152cbd4f012