common/rust/shed: build/test on same platforms as common/rust
Summary:
Before, `fbcode/common/rust/shed/` was not being built/tested on
macOS/Windows with citadel because of the `overwrite = True`, but now it
is.

With contbuild, `fbcode/common/rust/shed/**/facebook` was expected to
not work with `dev-rust-oss`, but now it does. We should be able to just
pass `...` and have things work.

Reviewed By: rjbailey

Differential Revision: D52080572

fbshipit-source-id: 79d898cf637bf4431477e6f804ff608de67f9455