Add From<{i128,u128}> for ScubaValue
Summary:
These were omitted because ScubaValue was landed in D4609607, a year before 128-bit integers became available in Rust 1.26.

It isn't correct to just cast these to i64 (as this macro will do after this diff), but it is not correct to just cast u64 or usize to i64 either (but this macro already does so).

Even if this isn't the right behavior, it's surely better than the current state of affairs, since whenever someone wants to log a duration to Scuba, [they just cast it to i64/u64 anyway](https://fburl.com/code/tvyqx5id). If the ideal behavior is to saturate/log an error/panic/encode as a Scuba normal when outside i64 range (or whatever else), we can at least change this From impl to do that later (and users of the From<[ui]128> impls will benefit from that change, unlike the casting callsites which would require a codemod).

Differential Revision: D53839755

fbshipit-source-id: 3b68813c70972ce91e6784438fcd55420472a4d1