Add support for skipping attributes in StructuredSample
Summary:
Previously if you wanted to derive `StructuredSample` on a struct
you would need to ensure that all fields were serializable and compatible
with `ScubaSample`.  Sometimes types cannot be easily serialized or the user
has a reason to omit them (such as privacy concerns).  This diff adds
the capability of skipping attributes on a struct that derives
`StructuredSample`.

Additionally fixed an issue where structs that define lifetimes were not
supported for deriving `StructuredSample`

Reviewed By: edward-shen

Differential Revision: D53949872

fbshipit-source-id: f35ef01c105d8e17882976ebd062bbee351fec82