Create TryFromSample derive macro
Summary:
Add a `TryFromSample` derive macro to facilitate converting between `ScubaSample` and custom structs. This is used in D53948478.

I chose not to add this functionality to `StructuredSample` because there are already some structs which use `StructuredSample` and have fields with custom `ScubaValue`s. It is not possible to create a generic way of implementing TryFrom for these custom fields. The new derive macro allows us to avoid breaking these structs. The two macros are compatible with each other if there is any use case where someone wants both functionalities.

Note: I refactor much of the logic in the existing derive macro to reuse it but didn't change this logic.

Reviewed By: kuecks

Differential Revision: D53964130

fbshipit-source-id: 191773de38b1df69f23ccfda0ce6c363c4314aae