Unify GenContext::Lib and GenContext::Types
Summary:
Since {D54710352}, there is no longer a separate "lib" crate and "types" crate.

I have chosen to keep the "types" terminology and get rid of "lib", rather than the other way around, because I find this a clearer way to talk about the crate containing the Thrift-generated types, even if there isn't a "-types" suffix on the crate name or Buck target.

Reviewed By: zertosh

Differential Revision: D54766547

fbshipit-source-id: e69f9bcf043de3d548e91e0d9dfdeaef072b5b9c