Add buffered_weighted crate for weight based concurrent polling of futures in a stream
Summary:
This crate essentially provides the buffered and ordered semantics of the standard [`buffered`](https://docs.rs/futures/latest/futures/stream/trait.StreamExt.html#method.buffered) combinator in futures crate but with an important distinction. The futures crate buffered variant treats every future the same (in terms of cost), but this crate allows each future to have a different weight. The constraint on the concurrency of polled futures is not the overall count of the futures, but the combined weight of the futures.

NOTE: This crate is heavily based off on the [`future-queue`](https://docs.rs/future-queue/latest/future_queue/index.html) written by Rain. Since it does not have support for ordered buffering, I have created this crate to make up for it. I will also try to upstream these changes directly in the `future-queue` crate at which point I will decomission this crate.

Differential Revision: D55965159

fbshipit-source-id: e07cafb3b352fbc4057a366c111fb0e170829281