Add a -mocks subcrate exporting constructor function for mock clients
Summary:
According to the analysis in D54709575, about 60% of the compile time of Rust Thrift client crates is the mocks. Mocks are rarely used compared to the rest of the contents of the clients crate, and almost always just in tests. Moving mocks out to a separate crate therefore makes clients build 2.5× faster.

This diff keeps the mock implementations in the `-clients` crate for now, but adds a new `-mocks` crate matching the API that we'll need to migrate to. According to `fbgs -frs$ '>::mock()' -l` there are 769 call sites that need to be updated, which is small compared to the Thrift codemods from earlier in the year. Once the codemod is done, we get to cash in on the build speedup in {D56451418}.

The dependency graph is:

```lang=mermaid
flowchart BT;
    foo-clients --> foo["foo (#quot;types#quot;)"];
    foo-services --> foo;
    foo-mocks --> foo-clients;
```

Previously, mocks were constructed using an associated function on the client trait object: `<dyn FooClient>::mock()`. In the new API, the trait `FooClient` is defined in `foo-clients` whereas the mock client for the same Thrift service is defined in `foo-mocks`, so a different approach is needed because the mocks crate cannot add an inherent method onto a trait object defined in a different crate. We need to pick one of the following APIs to use instead:

1. An extension trait: `use foo_mocks::DynClient as _; <dyn FooService>::mock()`

2. `<dyn FooService as foo_mocks::DynClient>::mock()`

3. `foo_mocks::new::<dyn FooService>()`

4. `foo_mocks::FooService::new()`

5. `foo_mocks::FooServiceMock::new()` (append `Mock` suffix to service name)

Number 3 or 5 seems the most promising to me. 4 doesn't seem pleasant because you'd almost always want to have the client `trait FooService` imported in order to be able to make trait method calls on your mock, which gives rise to a name collision in the type namespace.

This diff implements option 3 but I am interested in feedback.

Reviewed By: Imxset21, shayne-fletcher

Differential Revision: D56451421

fbshipit-source-id: c23fd9a28f8dd5bd75306a913819bc0ed66fb4e5