Simplify scuba_sample's generated TryFrom impls
Summary:
Someone got a bit sidetracked implementing this macro.

Part of it generates a whole "builder" type, with private fields, a `new` constructor, setters, and a fallible `build` method which does checking for whether all the fields were set.

But the whole thing is only used as a private implementation detail of 1 function, and that function always sets all the fields. So not only is the checking performed by `build` useless, but the whole builder is useless because we can just directly construct the desired output as a struct literal instead.

Reviewed By: diliop

Differential Revision: D59038432

fbshipit-source-id: 59d1d93d72156fcd80dc3f623cc6465a69362669