Ensure that `#[fbinit::test]` starts a mutli-threaded runtime (as of `#[finit::main]`)
Summary:
Extend `fbinit_test` to support configuration of created tokio runtime to run the tests on.

Depending on the presence of a value in `tokio_workers` we will create a `CurrentThread` or `MultiThread` runtime.

This is provided as `test` and `main`, without the change, would be running different runtimes and as such assumptions made on production code could not be really be verified.

Reviewed By: dtolnay

Differential Revision: D59221348

fbshipit-source-id: db741f47d5ac4d74923402d35d3d3652e82289e7