Update `sql_field` and `option_sql_field` macros to not make assumptions on imported symbols
Summary:
Before this diff the macros were expecting that
* `TryFromRowField` / `OptionalTryFromRowField` symbols exists on the macro caller's scope
  and that the symbol refers to the type from within the crate.
* `MysqlError` symbol exists on the macro caller's scope and that the symbol
  refers to `mysql_client::MysqlError`

This is an antipattern in macros definition, because the macro imposes the caller to do some imports that they might not need and imposes a certain semantic value of the symbol as well.

The reason why this is not great is that as a caller I might be using the macro in a context like
```
use mysql_client::TryFromRowField;
enum MysqlError {
    Random,
    SchemaError(String),
}
...
let field: String = sql_field!(row, Row, "field");
```
and the macro might or might not generate valid code.

To address this issues I'm looking into 2 changes:
* expected symbols defined within the library (`TryFromRowField` / `OptionalTryFromRowField`) will be prefixed with `$crate`.
  This effectivelly will make sure that the generated code is buildable.
* MysqlError on the other end is using the absolute import path such that at least the error will be clear on the absence of the `mysql_client` crate dependency.
  At the same time, the code will not make assumption on the content of `MysqlError` symbol

Differential Revision: D60018106

fbshipit-source-id: 97cc9e72615aa4b905e768bfb21ce1d4d5650e5f