Rewrite third-party helpers
Summary:
This diff rewrites the third party helpers to use constraints instead
of labels. This allows the necessary system packages to be resolved as part of
the query graph, and more reliably extracted than from labels. It also adds
implicit typo enforcement - since we are working with constraints, buck will
fail if a distro is typo'd.

This diff contains the new macros/rules in defs.bzl, rewrites the macros in
third_party.bzl to use them internally, and migrates a limited number of BUCK
files that depended on third_party_library() behavior that didn't work with the
macro rewrite. The remaining targets are migrated and third_party.bzl is
deleted in a stacked diff.

This diff depends on D63295525, which adds a new field to one of the prelude
helpers. The prelude is packaged with the buck2 binary, meaning if buck2 is run
through dotslash, it may be incompatible. If there are any errors related to a
"package" attribute, a newer version of buck2 is needed.

Reviewed By: bigfootjon

Differential Revision: D63653719

fbshipit-source-id: 2eb759300522feefdd59ff985056b2dacb53f322