Optimization: remove unnecessary string concatenation in ODS dynamic stats.
Summary:
Current macro does an extra join operation when prefix is available(most common case), this diff eliminates this operation to construct the key string in one shot eliminating a extra temp string allocation+free and the resulting fragmentation.

this operation is about 30% of the key generation (join + free):

https://fburl.com/strobelight/nr0vzg42

{F1974221233}

Motivated by the inefficiency from D67604234 in Re, but the app that will be benefit the most will be fb_monitord (a wdb) this optimization should account for ~0.23% CPU util improvement for fb_monitord. (the eliminated operation takes about 0.23% of the strobe-light samples)

__key_generator cost is currently around [40k $CPU-T1-equiv/year Q4 2024](https://fburl.com/strobelight/nk9vzdjd) so this is a 10k$/year optimization if I understand [the doc](https://www.internalfb.com/wiki/Capacity_Management_Internal/Capacity_Efficiency_pillar/Efficiency_Management_Framework/Measuring_Horizontal_CPU_Wins/) correctly.

Reviewed By: edward-shen

Differential Revision: D67913458

fbshipit-source-id: 08ddaa4de59f43744d202f4c71339580431290b6