facet: add build_from to build containers from other containers
Summary:
Add `Container::build_from` which allows a container to be built from another container, as long as the other container has at least the same facets as the container we are trying to build.

This replaces the `_from_container` macro, which didn't support delegates or fields, and ran into macro-generated-macros issues within the Rust language.

Differential Revision: D68207112

fbshipit-source-id: c3cf34caddfb37fdaa60fb84530e2d67873c68b2