futures_ext: remove slog dependency
Summary:
We are migrating from `slog` to `tracing`.

The `YieldPeriodically` stream adapter has the ability to log to a `slog::Logger`.  This is difficult to migrate to `tracing`, as it allows capturing of code locations which is not possible in tracing, as tracing uses static locations generated by its macros.

Instead, we remove the logging dependency and abstract it out to a callback that is called when polling exceeds the budged by a significant amount.  It is then up to the caller to log appropriately.  Since the logging happens in the caller, the log context, file name and line number are all correct.

Reviewed By: andreacampi

Differential Revision: D68837434

fbshipit-source-id: 6348754ec8cb2a45bec24b3022e4d715c4f6b561