add tracing_slog_compat
Summary:
We are migrating from `slog` to `tracing`.

To facilitate the migration, we add `tracing_slog_compat` which is a mostly-compatible drop-in replacement for the `slog` crate that can be used both with `slog` and `tracing` back-ends.

This crate is used by swapping out the `slog` crate dependency for a named dependency on this crate.

Not all features of `slog` are supported.  Some things are not simple to emulate, and some features are not supported by tracing anyway.  It primarily covers the slog logging macros.

Differential Revision: D68837436

fbshipit-source-id: 4780664919bfc91fe7cc12c988de354e5ffdf6ae