Upgrade prost to 0.12
Summary:
Upgrades `prost` to 0.12 and makes the related crates using `prost` compatible. I had to do some forking to make `tonic` happy, which I will resolve in D69623968.

Upgrading to 0.13 turns out to be a much bigger endeavor, as `tonic` upgrade will cause a world of deps to need to be upgraded as well (`tokio`, `hyper`, etc). Not sure how to handle that quite yet...

Reviewed By: GuyLewin, cryo28, dtolnay

Differential Revision: D69619522

fbshipit-source-id: 180c70131be7ea40b57928219c5905aa47857fc4