ctp: make SlogKVError take a reference to the error to be logged
Summary:
I am trying to get rid of `panic!()`s in the Multisect codebase. This
means I need to pass around errors more. There's a pattern of wrapping errors
in `SlogKVError` to use in logging with `slog`.

Previously `SlogKVError` took ownership of the error passed into it. This leads
to a bit of an akward pattern where an error might be logged, but then to get
it back for propagation up the stack it needs to be taken from the error struct
again.

Make `SlogKVError` take a reference to the error for logging. This is
sufficient for the logging use case.

Differential Revision: D75228264

fbshipit-source-id: ce30aff17af1f150d0cb8354bbe970e05e5483b6