Update cargo to 0.87.1
Summary:
This is challenging because `cargo` depends on `rusqlite` which depends on `libsqlite3-sys`. This diff updates libsqlite3-sys from 0.30.1 to 0.31.0. Both of these use `links = "sqlite3"`. So we must also atomically update every single other crate that has a dependency on libsqlite3-sys 0.30 to a version that depends on libsqlite3-sys 0.31. In addition to `cargo`, this includes:

- `r2d2_sqlite`
- `rusqlite`
- `rusqlite_migration`
- `serde_rusqlite`
- `sqlx`

For `r2d2_sqlite` and `rusqlite` and `serde_rusqlite` this is a simple dependency update. For `rusqlite_migration`, 1.x is no longer maintained so its dependency on libsqlite3-sys is not something that can be PR'd. For `sqlx`, their MSRV policy is incompatible with the newer libsqlite3-sys so it is also not something that can be PR'd.

Reviewed By: cjlongoria

Differential Revision: D75362932

fbshipit-source-id: b1e7e2d7004cfd474de740e4e73312f64f4c656b