add sample rate to Sampling.SampledOut
Summary:
The sample rate is still useful even when sampled out. For example when performing a sampling for another sample:

When a sample rate of 10 is applied and you want to log another sample with a sample rate of 20, you want to chanve that to a sample rate of 2 when you want to have both of those samples available. By making the sampled out rate available, this is easier to implement.

Reviewed By: kuecks

Differential Revision: D75680097

fbshipit-source-id: 4b0a33dd2f90c7686dd1af6a49799f56c0c79d7d