add support for setting env vars
Summary:
It normally only safe to call `std::env::set_var` before starting threads, but
`perform_init` actually starts one (`FuncSched`), so that should preferably
be done before doing init.

I need to do that in a Rust binary to set up `OMP_NUM_THREADS=1`, so now I have
the problem of having to stop using the fbinit macro.

But, not using the fbinit macro at all is a little annoying, because it already
has a bunch of logic that I'd like to reuse.

This diff proposes another solution, which is adding support in fbinit to set
up env vars. That seems like a pretty cheap thing to add, and it allows for a
pretty convenient way to set up runtime env vars in a way that's canonically
safe.

Thoughts?

Reviewed By: dtolnay

Differential Revision: D77221369

fbshipit-source-id: 680e664a20deeb1bdfaf3d3b1b42416c21199a84