Expose mysql test connection setup helper
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on testing.

## This diff
Extract the function that sets up a MySQL test connection to the tests_lib crate, so we can reuse it when we add the first functional unit tests in `mononoke_queries.rs` (D77935148).

Reviewed By: YousefSalama

Differential Revision: D77932383

fbshipit-source-id: 0f8660bc07b9506af037357f9ab7dfbe4f3b1cd2